+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to this project will be documented in this file.
 
  - Added spam `everyone`
 
+### Changed
+
+ - Better error management
+
 ## [0.0.2.1]
 
 ### Fixed